% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_pages.R
\name{help.pages}
\alias{help.pages}
\title{Locate URL links of Installed Packages}
\usage{
help.pages(package = full_stack, function_pattern = NULL,
  regex = FALSE, DT_args = list(options = list(paging = FALSE, dom =
  "t")))
}
\arguments{
\item{package}{character, names of pacakges to query, Default: full_stack}

\item{function_pattern}{character, pattern of functions to query within the
packages queried, Default: NULL}

\item{regex}{logical. If TRUE then package pattern is matched against the
list of installed package names, Default: FALSE}

\item{DT_args}{list, arguments to pass to datatable on print}
}
\value{
data.frame of class help_pages
}
\description{
Search for packages and functions for the URL documentation
 as listed in the package DESCRIPTION files
}
\details{
full_stack is created by the package on attachment/loading consists
 a vector of the installed packages on the machine.
}
\examples{
\dontrun{
if(interactive()){

  # return all packages with URLS
  help.pages()

  # return a specific subset of installed packages
  help.pages(package = c('tidyr','tidytext'))

  # search by partial match of pacakge names
  help.pages(package = 'tidy',regex = TRUE)

  # search by function pattern
  help.pages(function_pattern = 'geom')

 }
}
}
\seealso{
\code{\link[desc]{desc_get_urls}}
}
